[#12]; refactor: user domain
